requirements.txt
numpy
scipy
pandas
pyyaml
opencv-python
tqdm
ezc3d
rich
